Test finished, clear winner was AWS
I changed the json structure of the responses because I wanted to know the price.
I used another model version of Claude because the idea was to test from the cheapest to the expensive,
and I was wrong in choosing the cheapest models, since Amazon now shows good results even though it is
the cheapest of all, it would be used for the priming agent.
The Llama model is not the cheapest (from Llama) but it has prompt systems to test quickly,
since I can use more updated Llama models with better performance (and a little more expensive)
I did not want to do the test because it would take me time to structure the prompt systems as indicated in the documentation
(https://www.llama.com/docs/model-cards-and-prompt-formats/llama3_1#prompt-template).